+++ Cargo.toml
@@ -12,6 +12,7 @@ color-eyre = "0.6.2"
 figment = { version = "0.10.8", features = ["toml", "env"] }
 flexi_logger = "0.24.2"
 futures = "0.3"
+hex = "0.4"
 log = "0.4.17"
 prost-types = "0.11.8"
 rs-cnc = { git = "https://github.com/astriaorg/rs-cnc", rev = "e0b45d6" }
