--- src/main.rs
@@ -1,9 +1,0 @@
-use color_eyre::eyre::Result;
-
-use astria_conductor::run;
-
-#[tokio::main]
-async fn main() -> Result<()> {
-    run().await?;
-    Ok(())
-}
